+++ examples/states_ui/main.rs
@@ -1,0 +1,101 @@
+use amethyst::{
+    assets::HotReloadBundle,
+    audio::AudioBundle,
+    core::transform::TransformBundle,
+    input::{InputBundle, StringBindings},
+    prelude::*,
+    renderer::{
+        plugins::RenderToWindow,
+        // rendy::mesh::{Normal, Position, TexCoord},
+        types::DefaultBackend,
+        RenderingBundle,
+    },
+    ui::{RenderUi, UiBundle},
+    utils::application_root_dir,
+    utils::fps_counter::FpsCounterBundle,
+};
+
+mod credits;
+mod events;
+mod game;
+mod menu;
+mod pause;
+mod util;
+mod welcome;
+
+/// Quick overview what you can do when running this example:
+///
+/// Switch from the 'Welcome' Screen to the 'Menu' Screen.
+/// From the 'Menu', switch to either 'Credits' (from which you can only switch back) or to 'Game'.
+/// From 'Game', you can enter the 'Pause' menu.
+/// Here you can select to resume (go back to 'Game'), exit to menu (go to 'Menu') or exit (quit).
+///
+/// During the 'Pause' menu, the 'Game' is paused accordingly.
+
+pub fn main() -> amethyst::Result<()> {
+    amethyst::start_logger(Default::default());
+
+    // this will be the directory the 'Cargo.toml' is defined in.
+    let app_root = application_root_dir()?;
+
+    // our display config is in our configs folder.
+    let display_config_path = app_root.join("examples/states_ui/config/display.ron");
+
+    // other assets ('*.ron' files, '*.png' textures, '*.ogg' audio files, ui prefab files, ...) are here
+    let assets_dir = app_root.join("examples/assets");
+
+    let game_data = GameDataBuilder::default()
+        // a lot of other bundles/systems depend on this (without it being explicitly clear), so it
+        // makes sense to add it early on
+        .with_bundle(TransformBundle::new())?
+        // this bundle allows us to 'find' the Buttons and other UI elements later on
+        .with_bundle(UiBundle::<StringBindings>::new())?
+        // this allows us to reload '*.ron' files during execution
+        .with_bundle(HotReloadBundle::default())?
+        // without this Bundle, our Program will silently (!) fail when trying to start the 'Game'.
+        // (try it!)
+        // It takes care of Audio (in this case, the Button audio for hovering/clicking)
+        .with_bundle(AudioBundle::default())?
+        // With this System, we can register UI events and act accordingly.
+        // In this example it simply prints the events, excluding it does not provide less
+        // functionality.
+        .with_system_desc(
+            crate::events::UiEventHandlerSystemDesc::default(),
+            "ui_event_handler",
+            &[],
+        )
+        // This system is in 'events.rs'. Basically, it registers UI events that
+        // happen. Without it, the buttons will not react.
+        .with_bundle(InputBundle::<StringBindings>::new())?
+        // Necessary for the FPS counter in the upper left corner to work.
+        // (simply uncommenting will fail at runtime, since the resource is expected to exist, you
+        // need to uncomment line 107-114 in game.rs for it to still work)
+        .with_bundle(FpsCounterBundle)?
+        // Without this, we would not get a picture.
+        .with_bundle(
+            RenderingBundle::<DefaultBackend>::new()
+                // This creates the window and draws a background, if we don't specify a
+                // background in the loaded ui prefab file.
+                .with_plugin(
+                    RenderToWindow::from_config_path(display_config_path)?
+                        .with_clear([0.005, 0.005, 0.005, 1.0]),
+                )
+                // Without this, all of our beautiful UI would not get drawn.
+                // It will work, but we won't see a thing.
+                .with_plugin(RenderUi::default()),
+            // If you want to draw Sprites and such, you would need this additionally:
+            // .with_plugin(RenderFlat2D::default())
+        )?;
+
+    // creating the Application with the assets_dir, the first Screen, and the game_data with it's
+    // systems.
+    let mut game = Application::new(
+        assets_dir,
+        crate::welcome::WelcomeScreen::default(),
+        game_data,
+    )?;
+    log::info!("Starting with WelcomeScreen!");
+    game.run();
+
+    Ok(())
+}