+++ docs/CHANGELOG.md
@@ -11,10 +11,10 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ## [Unreleased]
 
 ### Added
-
 - Implement `Debug` for `ProgressCounter` and `ProgressCounterTracker`. ([#1973])
 - Added a custom render pass Example. ([#1904])
 - Add an entry for `examples/tiles` to the examples readme. ([#1978])
+- Added UI states/menu example. [#1986]
 
 ### Changed
 
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#1974]: https://github.com/amethyst/amethyst/pull/1974
 [#1978]: https://github.com/amethyst/amethyst/pull/1978
 [#1983]: https://github.com/amethyst/amethyst/pull/1983
+[#1986]: https://github.com/amethyst/amethyst/pull/1986
 [#1989]: https://github.com/amethyst/amethyst/pull/1989
 
 ## [0.13.3] - 2019-10-4